Extra commands and inactivity checking
This also goes in a CC response box. Select "Regex" as a trigger and type \A in the trigger box, just \A nothing else, as explained in the comments. This CC has no configuration variables, all you need to do is add it and select the correct trigger.
This will automatically close inactive tickets after a specified time (standard is 3h inactivity, this can be changed in the code which goes in the "Opening message in new tickets" box, which is explained in the next file